+++ src/page-content/models/article-interfaces.ts
